clean up cover page styling
This removes the header based changes that were in the main css file.
Instead of using element based markup, identifiers have been added for a
title and subtitle with proper offsets from the top. Also to address
the cover background, the cover section has been changed to white for
text. Finally, sample text is added for presenter name and date.